建築物のセーブ・ロードが完成

+++ Assets/Scripts/Building/BuildingInfo.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class BuildingInfo : MonoBehaviour
 {
     public int objID; // オブジェクトのID(どのゲームオブジェクトを使用しているか)
-    public ulong population; // 人口増減度
+    public int population; // 人口増減度
     public int cost; // 所持金増減度
     public int armamentsPower; // 軍備力増減度
     public int industryPower; // 工業力増減度
